stop-place-tile * Extract the toggle checkbox for routes to a component
