+++ src/containers/Admin/EditTab/StopPlaceTile/StopPlacePanel/PanelRow/RouteCheckbox.module.scss
@@ -1,0 +1,6 @@
+.Route {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    margin-bottom: 0.5rem;
+}
